+++ oxide_editor/Cargo.toml
@@ -9,6 +9,6 @@ repository.workspace = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-eframe = "0.19.0"
+eframe = "0.20.1"
 serde = { version = "1", features = ["derive"], optional = true }
 faux = "0.1"
